% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_builder}
\alias{xla_builder}
\title{Create an XLA computation builder}
\usage{
xla_builder(name = "computation")
}
\arguments{
\item{name}{A name for the computation (used in debug output).}
}
\value{
An external pointer of class \code{xla_builder}.
}
\description{
Create an XLA computation builder
}
