% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_devices}
\alias{xla_devices}
\title{List devices on a client}
\usage{
xla_devices(client)
}
\arguments{
\item{client}{An \code{xla_client} object.}
}
\value{
Character vector of device descriptions.
}
\description{
List devices on a client
}
