% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_dot}
\alias{xla_dot}
\title{Dot product / matrix multiply}
\usage{
xla_dot(lhs, rhs)
}
\arguments{
\item{lhs}{An \code{xla_op}.}

\item{rhs}{An \code{xla_op}.}
}
\value{
An \code{xla_op}.
}
\description{
Dot product / matrix multiply
}
