% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_mul}
\alias{xla_mul}
\title{Element-wise multiplication}
\usage{
xla_mul(lhs, rhs)
}
\arguments{
\item{lhs}{An \code{xla_op}.}

\item{rhs}{An \code{xla_op}.}
}
\value{
An \code{xla_op}.
}
\description{
Element-wise multiplication
}
