% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_reshape}
\alias{xla_reshape}
\title{Reshape}
\usage{
xla_reshape(x, new_dims)
}
\arguments{
\item{x}{An \code{xla_op}.}

\item{new_dims}{Integer vector of new dimensions.}
}
\value{
An \code{xla_op}.
}
\description{
Reshape
}
