% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_sub}
\alias{xla_sub}
\title{Element-wise subtraction}
\usage{
xla_sub(lhs, rhs)
}
\arguments{
\item{lhs}{An \code{xla_op}.}

\item{rhs}{An \code{xla_op}.}
}
\value{
An \code{xla_op}.
}
\description{
Element-wise subtraction
}
