% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_transpose}
\alias{xla_transpose}
\title{Transpose}
\usage{
xla_transpose(x, permutation)
}
\arguments{
\item{x}{An \code{xla_op}.}

\item{permutation}{Integer vector of dimension permutation (0-indexed).}
}
\value{
An \code{xla_op}.
}
\description{
Transpose
}
