% tinyrox says don't edit this manually, but it can't stop you!
\name{xla_client}
\alias{xla_client}
\title{Create an XLA client}
\usage{
xla_client()
}
\value{
An external pointer of class \code{xla_client}.
}
\description{
Creates a CPU-backed XLA client for compiling and executing computations.
}
